python/requirements.txt
phonenumbers
bottle
pydantic
waitress
zipcode